Separate spaceship and appstoreconnect api client functions.
