+++ autoprovision/certificatehelper.go
@@ -1,9 +1,7 @@
 package autoprovision
 
 import (
-	"crypto/x509"
 	"fmt"
-	"math/big"
 	"strings"
 
 	"github.com/bitrise-io/go-utils/log"
@@ -28,91 +26,6 @@ var CertificateTypeByDistribution = map[DistributionType]appstoreconnect.Certifi
 	AppStore:    appstoreconnect.IOSDistribution,
 	AdHoc:       appstoreconnect.IOSDistribution,
 	Enterprise:  appstoreconnect.IOSDistribution,
-}
-
-// APICertificateSource ...
-type APICertificateSource struct {
-	client *appstoreconnect.Client
-}
-
-// QueryCertificateBySerial ...
-func (s *APICertificateSource) QueryCertificateBySerial(serial *big.Int) (APICertificate, error) {
-	response, err := s.client.Provisioning.FetchCertificate(serial.Text(16))
-	if err != nil {
-		return APICertificate{}, err
-	}
-
-	certs, err := parseCertificatesResponse([]appstoreconnect.Certificate{response})
-	if err != nil {
-		return APICertificate{}, err
-	}
-	return certs[0], nil
-}
-
-// QueryAllIOSCertificates returns all iOS certificates from App Store Connect API
-func (s *APICertificateSource) QueryAllIOSCertificates() (map[appstoreconnect.CertificateType][]APICertificate, error) {
-	typeToCertificates := map[appstoreconnect.CertificateType][]APICertificate{}
-
-	for _, certType := range []appstoreconnect.CertificateType{appstoreconnect.Development, appstoreconnect.IOSDevelopment, appstoreconnect.Distribution, appstoreconnect.IOSDistribution} {
-		certs, err := queryCertificatesByType(s.client, certType)
-		if err != nil {
-			return map[appstoreconnect.CertificateType][]APICertificate{}, err
-		}
-		typeToCertificates[certType] = certs
-	}
-
-	return typeToCertificates, nil
-}
-
-// NewAPICertificateSource ...
-func NewAPICertificateSource(client *appstoreconnect.Client) CertificateSource {
-	return &APICertificateSource{
-		client: client,
-	}
-}
-
-func queryCertificatesByType(client *appstoreconnect.Client, certificateType appstoreconnect.CertificateType) ([]APICertificate, error) {
-	nextPageURL := ""
-	var certificates []appstoreconnect.Certificate
-	for {
-		response, err := client.Provisioning.ListCertificates(&appstoreconnect.ListCertificatesOptions{
-			PagingOptions: appstoreconnect.PagingOptions{
-				Limit: 20,
-				Next:  nextPageURL,
-			},
-			FilterCertificateType: certificateType,
-		})
-		if err != nil {
-			return nil, err
-		}
-		certificates = append(certificates, response.Data...)
-
-		nextPageURL = response.Links.Next
-		if nextPageURL == "" {
-			return parseCertificatesResponse(certificates)
-		}
-	}
-}
-
-func parseCertificatesResponse(response []appstoreconnect.Certificate) ([]APICertificate, error) {
-	var certifacteInfos []APICertificate
-	for _, resp := range response {
-		if resp.Type == "certificates" {
-			cert, err := x509.ParseCertificate(resp.Attributes.CertificateContent)
-			if err != nil {
-				return nil, fmt.Errorf("failed to parse certificate: %s", err)
-			}
-
-			certInfo := certificateutil.NewCertificateInfo(*cert, nil)
-
-			certifacteInfos = append(certifacteInfos, APICertificate{
-				Certificate: certInfo,
-				ID:          resp.ID,
-			})
-		}
-	}
-
-	return certifacteInfos, nil
 }
 
 // CertsToString ...
