+++ autoprovision/profilehelper.go
@@ -3,154 +3,16 @@ package autoprovision
 import (
 	"fmt"
 	"io/ioutil"
-	"net/http"
 	"os"
 	"path"
 	"strings"
 	"time"
 
-	"github.com/bitrise-io/go-utils/log"
 	"github.com/bitrise-io/go-utils/pathutil"
 	"github.com/bitrise-io/go-xcode/profileutil"
 	"github.com/bitrise-io/go-xcode/xcodeproject/serialized"
 	"github.com/bitrise-steplib/steps-ios-auto-provision-appstoreconnect/appstoreconnect"
 )
-
-// Profile ...
-type Profile interface {
-	ID() string
-	Attributes() appstoreconnect.ProfileAttributes
-	CertificateIDs() (map[string]bool, error)
-	DeviceIDs() (map[string]bool, error)
-	BundleID() (appstoreconnect.BundleID, error)
-}
-
-// ProfileClient ...
-type ProfileClient interface {
-	FindProfile(name string, profileType appstoreconnect.ProfileType) (Profile, error)
-	DeleteProfile(id string) error
-	CreateProfile(name string, profileType appstoreconnect.ProfileType, bundleID appstoreconnect.BundleID, certificateIDs []string, deviceIDs []string) (Profile, error)
-	// Bundle ID
-	FindBundleID(bundleIDIdentifier string) (*appstoreconnect.BundleID, error)
-	CheckBundleIDEntitlements(bundleID appstoreconnect.BundleID, projectEntitlements Entitlement) error
-	SyncBundleID(bundleID appstoreconnect.BundleID, entitlements Entitlement) error
-	CreateBundleID(bundleIDIdentifier string) (*appstoreconnect.BundleID, error)
-}
-
-// APIProfile ...
-type APIProfile struct {
-	profile *appstoreconnect.Profile
-	client  *appstoreconnect.Client
-}
-
-// NewAPIProfile ...
-func NewAPIProfile(client *appstoreconnect.Client, profile *appstoreconnect.Profile) Profile {
-	return &APIProfile{
-		profile: profile,
-		client:  client,
-	}
-}
-
-// ID ...
-func (p APIProfile) ID() string {
-	return p.profile.ID
-}
-
-// Attributes ...
-func (p APIProfile) Attributes() appstoreconnect.ProfileAttributes {
-	return p.profile.Attributes
-}
-
-// CertificateIDs ...
-func (p APIProfile) CertificateIDs() (map[string]bool, error) {
-	var nextPageURL string
-	var certificates []appstoreconnect.Certificate
-	for {
-		response, err := p.client.Provisioning.Certificates(
-			p.profile.Relationships.Certificates.Links.Related,
-			&appstoreconnect.PagingOptions{
-				Limit: 20,
-				Next:  nextPageURL,
-			},
-		)
-		if err != nil {
-			return nil, wrapInProfileError(err)
-		}
-
-		certificates = append(certificates, response.Data...)
-
-		nextPageURL = response.Links.Next
-		if nextPageURL == "" {
-			break
-		}
-	}
-
-	ids := map[string]bool{}
-	for _, cert := range certificates {
-		ids[cert.ID] = true
-	}
-
-	return ids, nil
-}
-
-// DeviceIDs ...
-func (p APIProfile) DeviceIDs() (map[string]bool, error) {
-	var nextPageURL string
-	ids := map[string]bool{}
-	for {
-		response, err := p.client.Provisioning.Devices(
-			p.profile.Relationships.Devices.Links.Related,
-			&appstoreconnect.PagingOptions{
-				Limit: 20,
-				Next:  nextPageURL,
-			},
-		)
-		if err != nil {
-			return nil, wrapInProfileError(err)
-		}
-
-		for _, dev := range response.Data {
-			ids[dev.ID] = true
-		}
-
-		nextPageURL = response.Links.Next
-		if nextPageURL == "" {
-			break
-		}
-	}
-
-	return ids, nil
-}
-
-// BundleID ...
-func (p APIProfile) BundleID() (appstoreconnect.BundleID, error) {
-	bundleIDresp, err := p.client.Provisioning.BundleID(p.profile.Relationships.BundleID.Links.Related)
-	if err != nil {
-		return appstoreconnect.BundleID{}, err
-	}
-
-	return bundleIDresp.Data, nil
-}
-
-// APIProfileClient ...
-type APIProfileClient struct {
-	client *appstoreconnect.Client
-}
-
-// NewAPIProfileClient ...
-func NewAPIProfileClient(client *appstoreconnect.Client) ProfileClient {
-	return &APIProfileClient{client: client}
-}
-
-// NonmatchingProfileError is returned when a profile/bundle ID does not match project requirements
-// It is not a fatal error, as the profile can be regenerated
-type NonmatchingProfileError struct {
-	Reason string
-}
-
-func (e NonmatchingProfileError) Error() string {
-	return fmt.Sprintf("provisioning profile does not match requirements: %s", e.Reason)
-}
 
 // ProfileName generates profile name with layout: Bitrise <platform> <distribution type> - (<bundle id>)
 func ProfileName(profileType appstoreconnect.ProfileType, bundleID string) (string, error) {
@@ -172,72 +34,6 @@ func ProfileName(profileType appstoreconnect.ProfileType, bundleID string) (stri
 	}
 
 	return fmt.Sprintf("%sBitrise %s %s - (%s)", prefix, platform, distribution, bundleID), nil
-}
-
-// FindProfile ...
-func (c *APIProfileClient) FindProfile(name string, profileType appstoreconnect.ProfileType) (Profile, error) {
-	opt := &appstoreconnect.ListProfilesOptions{
-		PagingOptions: appstoreconnect.PagingOptions{
-			Limit: 1,
-		},
-		FilterProfileType: profileType,
-		FilterName:        name,
-	}
-
-	r, err := c.client.Provisioning.ListProfiles(opt)
-	if err != nil {
-		return nil, err
-	}
-	if len(r.Data) == 0 {
-		return nil, nil
-	}
-
-	return NewAPIProfile(c.client, &r.Data[0]), nil
-}
-
-func (c *APIProfileClient) deleteExpiredProfile(bundleID *appstoreconnect.BundleID, profileName string) error {
-	var nextPageURL string
-	var profile *appstoreconnect.Profile
-
-	for {
-		response, err := c.client.Provisioning.Profiles(bundleID.Relationships.Profiles.Links.Related, &appstoreconnect.PagingOptions{
-			Limit: 20,
-			Next:  nextPageURL,
-		})
-		if err != nil {
-			return err
-		}
-
-		for _, d := range response.Data {
-			if d.Attributes.Name == profileName {
-				profile = &d
-				break
-			}
-		}
-
-		nextPageURL = response.Links.Next
-		if nextPageURL == "" {
-			break
-		}
-	}
-
-	if profile == nil {
-		return fmt.Errorf("failed to find profile: %s", profileName)
-	}
-
-	return c.DeleteProfile(profile.ID)
-}
-
-func wrapInProfileError(err error) error {
-	if respErr, ok := err.(appstoreconnect.ErrorResponse); ok {
-		if respErr.Response != nil && respErr.Response.StatusCode == http.StatusNotFound {
-			return NonmatchingProfileError{
-				Reason: fmt.Sprintf("profile was concurrently removed from Developer Portal: %v", err),
-			}
-		}
-	}
-
-	return err
 }
 
 func checkProfileEntitlements(client ProfileClient, prof Profile, projectEntitlements Entitlement) error {
@@ -374,70 +170,6 @@ func CheckProfile(client ProfileClient, prof Profile, entitlements Entitlement, 
 		return err
 	}
 	return checkProfileDevices(profileDeviceIDs, deviceIDs)
-}
-
-// DeleteProfile ...
-func (c *APIProfileClient) DeleteProfile(id string) error {
-	if err := c.client.Provisioning.DeleteProfile(id); err != nil {
-		if respErr, ok := err.(appstoreconnect.ErrorResponse); ok {
-			if respErr.Response != nil && respErr.Response.StatusCode == http.StatusNotFound {
-				return nil
-			}
-		}
-
-		return err
-	}
-
-	return nil
-}
-
-// CreateProfile ...
-func (c *APIProfileClient) CreateProfile(name string, profileType appstoreconnect.ProfileType, bundleID appstoreconnect.BundleID, certificateIDs []string, deviceIDs []string) (Profile, error) {
-	profile, err := c.createProfile(name, profileType, bundleID, certificateIDs, deviceIDs)
-	if err != nil {
-		// Expired profiles are not listed via profiles endpoint,
-		// so we can not catch if the profile already exist but expired, before we attempt to create one with the managed profile name.
-		// As a workaround we use the BundleID profiles relationship url to find and delete the expired profile.
-		if isMultipleProfileErr(err) {
-			log.Warnf("  Profile already exists, but expired, cleaning up...")
-			if err := c.deleteExpiredProfile(&bundleID, name); err != nil {
-				return nil, fmt.Errorf("expired profile cleanup failed: %s", err)
-			}
-
-			profile, err = c.createProfile(name, profileType, bundleID, certificateIDs, deviceIDs)
-			if err != nil {
-				return nil, err
-			}
-
-			return profile, nil
-		}
-
-		return nil, err
-	}
-
-	return profile, nil
-}
-
-func (c *APIProfileClient) createProfile(name string, profileType appstoreconnect.ProfileType, bundleID appstoreconnect.BundleID, certificateIDs []string, deviceIDs []string) (Profile, error) {
-	// Create new Bitrise profile on App Store Connect
-	r, err := c.client.Provisioning.CreateProfile(
-		appstoreconnect.NewProfileCreateRequest(
-			profileType,
-			name,
-			bundleID.ID,
-			certificateIDs,
-			deviceIDs,
-		),
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create %s provisioning profile for %s bundle ID: %s", profileType.ReadableString(), bundleID.Attributes.Identifier, err)
-	}
-
-	return NewAPIProfile(c.client, &r.Data), nil
-}
-
-func isMultipleProfileErr(err error) bool {
-	return strings.Contains(strings.ToLower(err.Error()), "multiple profiles found with the name")
 }
 
 // WriteProfile writes the provided profile under the `$HOME/Library/MobileDevice/Provisioning Profiles` directory.
